requirements.txt
matplotlib==3.6.3
RPi.GPIO==0.7.0

